Use "mask" field for request functions
The XML can provide an "enum" property on a <field> of a request. This
indicates that the field is one of the valid types of that enum. We
already use this to expect an instance of that enum for the request.

There is also a "mask" property meaning that the field is a
logical-or-combination of values from the enum that the property refers
to. So far, we ignored this property. This commit changes the generated
code to expect an instance of the corresponding enum.

This change is possible since (for a while already) we no longer
consider enums to be exhaustive. Instead, they are just a newtype around
some integer and thus invalid enum values can also be represented.

+++ x11rb/examples/simple_window_manager.rs
@@ -339,7 +339,7 @@ impl<'a, C: Connection> WmState<'a, C> {
     }
 
     fn handle_button_press(&mut self, event: ButtonPressEvent) {
-        if event.detail != DRAG_BUTTON || event.state != 0 {
+        if event.detail != DRAG_BUTTON || u16::from(event.state) != 0 {
             return;
         }
         if let Some(state) = self.find_window_by_id(event.event) {
